Move estimate_without_restriction in own module.
